requirements.txt
# Data processing and database
pandas>=1.5.0
openpyxl>=3.0.0
sqlalchemy>=1.4.0

# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
openai>=1.0.0

# FastAPI and web framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0

# Security and rate limiting
slowapi>=0.1.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Environment and configuration
python-dotenv>=1.0.0

# Logging and monitoring
structlog>=23.0.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

# Development tools (optional)
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
